# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SelahKeep — Private Encrypted Notes API
  version: 0.1.0
  description: |
    REST API for the SelahKeep private encrypted notes MVP.

    **Security:** Private letter title and body must never be sent in plaintext.
    All letter fields use structured encrypted payloads (see ADR-001 / ADR-003).

    **Auth:** Most routes require a NextAuth session cookie (sign in via the web app first).
    Use browser Swagger "Try it out" while logged in, or import the spec into Postman/Insomnia
    with your session cookie.

    **Spec file:** `docs/openapi.yaml` — served as JSON at `GET /api/openapi`.
servers:
  - url: http://localhost:3001
    description: Local development (default port)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth JWT session cookie. Name may be `__Secure-next-auth.session-token` in production HTTPS.
    integrationBearer:
      type: http
      scheme: bearer
      bearerFormat: sk_int
      description: |
        User integration token (`sk_int_...`) for MCP routes. Created once at POST /api/integrations.
        Requires INTEGRATIONS_ENABLED=true.
...
paths:
...
  /api/passkeys/authenticate:
    post:
      tags: [Passkeys]
      summary: Vault unlock routing (options, sanitized verify candidates, or post-match bind)
      security:
        - sessionCookie: []
      responses:
        '200':
          description: WebAuthn options, encrypted candidates, or browser binding result; never PRF output
        '409':
          description: Stale browser binding cleared; explicit passkey selection/rebind required
    delete:
      tags: [Passkeys]
      summary: Unbind this browser only; credential and variants remain active
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Browser binding removed and cookie cleared
...
  /api/account/passkeys/{id}/vault-unlock:
    get:
      tags: [Account]
      summary: Get vault capability and active variant count for a passkey
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Vault passkey status }
    delete:
      tags: [Account]
      summary: Disable credential vault capability after local candidate match
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: All variants revoked and all browser bindings removed atomically }
...